ifcfg-rh: allow commenting values in ifcfg variables (rh #1369380)
This will allow placing a comment after values, like this:
  ONBOOT=no  # disable autoconnect
If a comment is inside quotes, it is preserved:
  NAME="My connection #1"

Notes:
1) Removing comments means a slight behaviour change, because
  NAME=My connection #1
returned 'My connection #1' string before, but now it will return
'My connection'. However, this is acceptable, because arguably a string with
spaces should have been quoted to preserve the content.

2) We remove comments even if the value is read verbatim, so that placing
comments would work for all ifcfg variables. Thus the actual value has to be
protected with quotes if it contains ' #' string resembling a comment.
Anyway, ifcfg reader removes trailing spaces even for verbatim read values,
so the quotes are necessary for preserving the spaces in any case.

https://bugzilla.redhat.com/show_bug.cgi?id=1369380